+++ apps/indexer/lib/indexer/fetcher/polygon_edge/withdrawal.ex
@@ -12,7 +12,7 @@ defmodule Indexer.Fetcher.PolygonEdge.Withdrawal do
 
   import EthereumJSONRPC, only: [quantity_to_integer: 1]
   import Explorer.Helper, only: [decode_data: 2]
-  import Indexer.Fetcher.PolygonEdge, only: [fill_block_range: 5, get_block_number_by_tag: 3]
+  import Indexer.Fetcher.PolygonEdge, only: [fill_block_range: 5]
   import Indexer.Helper, only: [log_topic_to_string: 1]
 
   alias ABI.TypeDecoder
@@ -20,6 +20,7 @@ defmodule Indexer.Fetcher.PolygonEdge.Withdrawal do
   alias Explorer.Chain.Log
   alias Explorer.Chain.PolygonEdge.Withdrawal
   alias Indexer.Fetcher.PolygonEdge
+  alias Indexer.Helper
 
   @fetcher_name :polygon_edge_withdrawal
 
@@ -107,7 +108,7 @@ defmodule Indexer.Fetcher.PolygonEdge.Withdrawal do
 
     if not safe_block_is_latest do
       # find and fill all events between "safe" and "latest" block (excluding "safe")
-      {:ok, latest_block} = get_block_number_by_tag("latest", json_rpc_named_arguments, 100_000_000)
+      {:ok, latest_block} = Helper.get_block_number_by_tag("latest", json_rpc_named_arguments, 100_000_000)
 
       fill_block_range(
         safe_block + 1,
