+++ apps/indexer/lib/indexer/helper.ex
@@ -3,6 +3,17 @@ defmodule Indexer.Helper do
   Auxiliary common functions for indexers.
   """
 
+  require Logger
+
+  import EthereumJSONRPC,
+    only: [
+      fetch_block_number_by_tag: 2,
+      json_rpc: 2,
+      quantity_to_integer: 1,
+      request: 1
+    ]
+
+  alias EthereumJSONRPC.Block.ByNumber
   alias Explorer.Chain.Hash
 
   @spec address_hash_to_string(binary(), boolean()) :: binary()
@@ -31,6 +42,113 @@ defmodule Indexer.Helper do
 
   def address_correct?(_address) do
     false
+  end
+
+  @spec get_block_number_by_tag(binary(), list(), integer()) :: {:ok, non_neg_integer()} | {:error, atom()}
+  def get_block_number_by_tag(tag, json_rpc_named_arguments, retries \\ 3) do
+    error_message = &"Cannot fetch #{tag} block number. Error: #{inspect(&1)}"
+    repeated_call(&fetch_block_number_by_tag/2, [tag, json_rpc_named_arguments], error_message, retries)
+  end
+
+  def get_transaction_by_hash(hash, json_rpc_named_arguments, retries_left \\ 3)
+
+  def get_transaction_by_hash(hash, _json_rpc_named_arguments, _retries_left) when is_nil(hash), do: {:ok, nil}
+
+  def get_transaction_by_hash(hash, json_rpc_named_arguments, retries) do
+    req =
+      request(%{
+        id: 0,
+        method: "eth_getTransactionByHash",
+        params: [hash]
+      })
+
+    error_message = &"eth_getTransactionByHash failed. Error: #{inspect(&1)}"
+
+    repeated_call(&json_rpc/2, [req, json_rpc_named_arguments], error_message, retries)
+  end
+
+  def log_blocks_chunk_handling(chunk_start, chunk_end, start_block, end_block, items_count, layer) do
+    is_start = is_nil(items_count)
+
+    {type, found} =
+      if is_start do
+        {"Start", ""}
+      else
+        {"Finish", " Found #{items_count}."}
+      end
+
+    target_range =
+      if chunk_start != start_block or chunk_end != end_block do
+        progress =
+          if is_start do
+            ""
+          else
+            percentage =
+              (chunk_end - start_block + 1)
+              |> Decimal.div(end_block - start_block + 1)
+              |> Decimal.mult(100)
+              |> Decimal.round(2)
+              |> Decimal.to_string()
+
+            " Progress: #{percentage}%"
+          end
+
+        " Target range: #{start_block}..#{end_block}.#{progress}"
+      else
+        ""
+      end
+
+    if chunk_start == chunk_end do
+      Logger.info("#{type} handling #{layer} block ##{chunk_start}.#{found}#{target_range}")
+    else
+      Logger.info("#{type} handling #{layer} block range #{chunk_start}..#{chunk_end}.#{found}#{target_range}")
+    end
+  end
+
+  def repeated_call(func, args, error_message, retries_left) do
+    case apply(func, args) do
+      {:ok, _} = res ->
+        res
+
+      {:error, message} = err ->
+        retries_left = retries_left - 1
+
+        if retries_left <= 0 do
+          Logger.error(error_message.(message))
+          err
+        else
+          Logger.error("#{error_message.(message)} Retrying...")
+          :timer.sleep(3000)
+          repeated_call(func, args, error_message, retries_left)
+        end
+    end
+  end
+
+  def get_block_timestamp_by_number(number, json_rpc_named_arguments, retries \\ 3) do
+    func = &get_block_timestamp_by_number_inner/2
+    args = [number, json_rpc_named_arguments]
+    error_message = &"Cannot fetch block ##{number} or its timestamp. Error: #{inspect(&1)}"
+    repeated_call(func, args, error_message, retries)
+  end
+
+  defp get_block_timestamp_by_number_inner(number, json_rpc_named_arguments) do
+    result =
+      %{id: 0, number: number}
+      |> ByNumber.request(false)
+      |> json_rpc(json_rpc_named_arguments)
+
+    with {:ok, block} <- result,
+         false <- is_nil(block),
+         timestamp <- Map.get(block, "timestamp"),
+         false <- is_nil(timestamp) do
+      {:ok, quantity_to_integer(timestamp)}
+    else
+      {:error, message} ->
+        {:error, message}
+
+      true ->
+        {:error, "RPC returned nil."}
+    end
   end
 
   @spec log_topic_to_string(any()) :: binary() | nil
