+++ apps/indexer/lib/indexer/transform/shibarium/bridge.ex
@@ -7,6 +7,8 @@ defmodule Indexer.Transform.Shibarium.Bridge do
 
   import Explorer.Chain.SmartContract, only: [burn_address_hash_string: 0]
 
+  import Indexer.Fetcher.Shibarium.Helper, only: [prepare_insert_items: 2]
+
   import Indexer.Fetcher.Shibarium.L2, only: [withdraw_method_signature: 0]
 
   alias Indexer.Fetcher.Shibarium.L2
@@ -26,9 +28,9 @@ defmodule Indexer.Transform.Shibarium.Bridge do
            child_chain = Application.get_env(:indexer, Indexer.Fetcher.Shibarium.L2)[:child_chain],
            weth = Application.get_env(:indexer, Indexer.Fetcher.Shibarium.L2)[:weth],
            bone_withdraw = Application.get_env(:indexer, Indexer.Fetcher.Shibarium.L2)[:bone_withdraw],
-           true <- Helper.is_address_correct?(child_chain),
-           true <- Helper.is_address_correct?(weth),
-           true <- Helper.is_address_correct?(bone_withdraw) do
+           true <- Helper.address_correct?(child_chain),
+           true <- Helper.address_correct?(weth),
+           true <- Helper.address_correct?(bone_withdraw) do
         child_chain = String.downcase(child_chain)
         weth = String.downcase(weth)
         bone_withdraw = String.downcase(bone_withdraw)
@@ -37,7 +39,7 @@ defmodule Indexer.Transform.Shibarium.Bridge do
         start_block = Enum.min(block_numbers)
         end_block = Enum.max(block_numbers)
 
-        L2.log_blocks_chunk_handling(start_block, end_block, start_block, end_block, nil, "L2")
+        Helper.log_blocks_chunk_handling(start_block, end_block, start_block, end_block, nil, "L2")
 
         deposit_transaction_hashes =
           transactions_with_receipts
@@ -66,9 +68,9 @@ defmodule Indexer.Transform.Shibarium.Bridge do
         timestamps = Enum.reduce(blocks, %{}, fn block, acc -> Map.put(acc, block.number, block.timestamp) end)
 
         operations = L2.prepare_operations({events, timestamps}, weth)
-        items = L2.prepare_insert_items(operations)
+        items = prepare_insert_items(operations, L2)
 
-        L2.log_blocks_chunk_handling(
+        Helper.log_blocks_chunk_handling(
           start_block,
           end_block,
           start_block,
